requirements.txt
Pillow
scipy
tensorflow
pytest-cov
pyobjc
flask
py2app
